% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_nodes}
\alias{add_nodes}
\title{Add nodes to neo4j from a dataframe}
\usage{
add_nodes(df, label, id, props = NULL, verbose = FALSE, na.act = F)
}
\arguments{
\item{df:}{dataframe which will be traversed to row-wise to add nodes.}

\item{label:}{The label to assign to each node (takes a character value)}

\item{id:}{column name in dataframe required for a node to be created (takes a character value)}

\item{props:}{column names(s) in dataframe of node properties (takes a character vector); defaults to NULL}

\item{verbose:}{if TRUE, returns information about each node being added to the graph DB; defaults to False}

\item{na.act:}{not currently used...maybe in the future; defaults to False}
}
\description{
This function adds nodes row-wise to a neo4j graph (called "graph")
}
\examples{
add_nodes(awarded_contracts, "contract", "DUNSNumber", props = names(awarded_contracts), verbose = T)
}
\keyword{nodes}

