% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rneo4_functions.R
\name{add_or_update_nodes}
\alias{add_or_update_nodes}
\title{Add or Update nodes in neo4j from a dataframe}
\usage{
add_or_update_nodes(df, label, id, props = NULL, verbose = FALSE, na.act = F)
}
\arguments{
\item{label:}{The label to assign to each node (takes a character value)}

\item{id:}{column name in dataframe required for a node to be created (takes a character value)}

\item{props:}{column names(s) in dataframe of node properties (takes a character vector); defaults to NULL}

\item{verbose:}{if TRUE, returns information about each node being added to the graph DB; defaults to False}

\item{na.act:}{not currently used...maybe in the future; defaults to False}
}
\description{
This function Checks to see if a node already exists with the label/id combination; if not, it adds the node to Neo4j; if so, it attempts update the node based on the properties in the row. The neo4j graph must be called "graph".  In order for the update portion to work, a constraint must be set on the label/id combination. Example: addConstraint(graph, "vendor", "DUNSNumber")
}
\examples{
add_or_update_nodes(awarded_contracts, "contract", "DUNSNumber", props = names(awarded_contracts), verbose = T)
}
\keyword{nodes}

