% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{convert_facts_to_chars}
\alias{convert_facts_to_chars}
\title{converts all factor variables in a data.frame to character variables}
\usage{
convert_facts_to_chars(df)
}
\arguments{
\item{df:}{the data.frame to have its factor columns converted to characters}
}
\description{
This function returns the original data.frame with all factor variables converted to character variables
}
\examples{
fake_df <- data.frame(alpha = as.factor(c('a', 'b', 'c', 'd')), beta = 1:4)
str(fake_df)
str(convert_facts_to_chars(fake_df))
}
\keyword{conversion}
\keyword{type}

