% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outlier_detection_functions.R
\name{dbscan_outlier}
\alias{dbscan_outlier}
\title{This function Calculates outliers using the DBSCAN algorithm}
\usage{
dbscan_outlier(df, eps_range = seq(0.5, 2, 0.3), MinPts_range = seq(3, 6, 1), scale = T, skip = c())
}
\arguments{
\item{df:}{data.frame with features over which to run the algorithm}

\item{eps_range:}{range of eps values to test in order to provide feedback to the user about what the algorithm is finding}

\item{MinPts_range:}{range of eps values to test in order to provide feedback to the user about what the algorithm is finding}

\item{scale:}{whether or not to scale the data prior to running the algorithm}

\item{skip:}{character vector of names of columns to be skipped}
}
\description{
Given a data.frame, returns a vector indicating whether each observations was deemed an outlier or not
}

