% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{dt_left_outer}
\alias{dt_left_outer}
\title{Perform left outer join of data.tables with diagnostics}
\usage{
dt_left_outer(dt1, dt2, key1, key2 = key1, dt2_subset_cols = NULL, cartesian = T)
}
\arguments{
\item{dt1:}{First data.table (the one "on the left")}

\item{dt2:}{Second data.table (the one "on the right")}

\item{key1:}{The column (character) from dt1 on which the join is being performed}

\item{key2:}{The column (character) from dt2 on which the join is being performed}

\item{dt2_subset_cols:}{Character vector of the columns to keep in dt2 after the join is performed}
}
\description{
Given two data.tables, this function performs a left outer join and reports on any cartesian product, what percentage matched up, etc.
}
\examples{
first_dt <- data.table (a = 1:5, b = c('a', 'b', 'c', 'd', 'e'))
second_dt <- data.table (a = 3:7, c = c('yes', 'no', 'no', 'yes', 'yes'))
dt_left_outer (first_dt, second_dt, 'a')
}
\keyword{database}
\keyword{exploration}

