% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{eliminate_single_value_cols}
\alias{eliminate_single_value_cols}
\title{Subsets to columns with more than one unique value}
\usage{
eliminate_single_value_cols(df)
}
\arguments{
\item{df:}{the data.frame to have its single-value columns removed}
}
\description{
This function returns the original data.frame without columns that only have one value for all rows in the data.frame
}
\examples{
fake_df <- data.frame(a = rep(1, 5), b = (1:5), c = rep('a', 5), d = c('a', 'b', 'c', 'd', 'e'))
eliminate_single_value_cols(fake_df)
}

