% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{find_unique_id}
\alias{find_unique_id}
\title{Find columns that uniquely identify a row in a dataframe}
\usage{
find_unique_id(df)
}
\arguments{
\item{df:}{dataframe}
}
\description{
This function returns the name of each column which uniquely identifies a row in the data.frame
}
\examples{
fake_df <- data.frame(a = 1:10, b = rep(1:5, 2))
find_unique_id(fake_df)
}
\keyword{id}
\keyword{key,}
\keyword{primary}
\keyword{unique}

