% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{get_all_table_info}
\alias{get_all_table_info}
\title{Get all table, column, and view metadata from SQL Server database}
\usage{
get_all_table_info(conn, table_names)
}
\arguments{
\item{conn:}{odbc handle}
}
\description{
Given an odbc handle this function gras all the metadata associated with a database at the table/column level and returns it as a data.frame
}
\examples{
get_all_table_info(conn)
}
\keyword{database}
\keyword{exploration}

