% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{get_table_lengths}
\alias{get_table_lengths}
\title{Gets the number of rows for a list of tables}
\usage{
get_table_lengths(conn, table_names)
}
\arguments{
\item{conn:}{odbc handle}

\item{table_names:}{character vector indicating the names of tables in the database for which the number of rows is desired}
}
\description{
Given an odbc handle and a vector of table names, this function returns a named vector containing the number rows in each table
}
\examples{
get_table_lengths(conn, c('best_table', 'next_best_table'))
}
\keyword{database}
\keyword{exploration}

