% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{get_table_sample}
\alias{get_table_sample}
\title{Take a sample of rows from a table in a database}
\usage{
get_table_sample(conn, table_name, samp_num)
}
\arguments{
\item{conn:}{ODBC database connection handle}

\item{table_name:}{character vector indicating the name of the table to be sampled from}

\item{samp_num:}{integer indicating the number of rows to sample}
}
\description{
This function returns a data.frame corresponding to rows in a database table that have been (somewhat) randomly sampled.
}
\examples{
get_table_sample(conn, 'jerrys_table', 15) # for sampling 15 rows from jerrys_table
}
\keyword{database}
\keyword{sampling,}

