% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{get_unique_ids}
\alias{get_unique_ids}
\title{Gets all columns that are unique identifiers from a list of tables}
\usage{
get_unique_ids(table_char_vect, conn, samp_num)
}
\arguments{
\item{table_char_vect:}{a character vector containing the table names to inspect for unique columns}

\item{conn:}{ODBC database handle}

\item{samp_num:}{integer indicating the number of rows to sample}
}
\description{
This function returns a data.frame that gives all unique columns found in the tables provided from a given database
}
\examples{
get_unique_ids(c('jerrys_table', 'jans_table', conn, 1000) # for obtaining unique identifiers from jan and jerrys table by checking 1000 randomly sample rows from each table
}
\keyword{database}
\keyword{sampling,}

