% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ip_functions.R
\name{is_local_ip}
\alias{is_local_ip}
\title{Determine whether or not an IP address is local.}
\usage{
is_local_ip(ip_vect)
}
\arguments{
\item{ip_vect:}{character vector containing ip addresses}
}
\description{
Given a character vector of IP addresses, returns a logical vector indicating whether each entry is local (T) or not (F)
}
\examples{
fake_ip_vect <- c('192.168.1.101', '10.10.2.201', '42.1.1.1', '125.64.3.111')
is_local_ip(fake_ip_vect)
}
\keyword{ip,}
\keyword{mask}

