% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outlier_detection_functions.R
\name{kmeans_outlier}
\alias{kmeans_outlier}
\title{Detects outliers using k-means clustering}
\usage{
kmeans_outlier (df, cluster_max = 10, percentile = 95, k_scale = T, skip = c())
}
\arguments{
\item{df:}{data.frame over which to run the algorithm}

\item{cluster_max:}{the maximum number of clusters to test for providing feedback to the user}

\item{percentile:}{the distance percentile in order to be deemed an outlier?}

\item{scale:}{whether or not to scale the inputs}

\item{skip:}{characer vector of which columns the algorithm should skip}
}
\description{
Given a data.frame, determines which points lie furthest away from their cluster centroids
}

