% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph_functions.R
\name{make_edges_dt}
\alias{make_edges_dt}
\title{Create a data.table that can be used to load all edges into the graph}
\usage{
make_edges_dt(dt, edge_info, same_e_attr_list = NULL)
}
\arguments{
\item{dt:}{data.table containing columns for all the edges and edge attributes}

\item{edge_info:}{named list, where the name is the edge type and each "to" and "from" refers to the columns in the dt to go from and to}

\item{same_e_attr_list:}{named list where the name is the edge attributes to resolve and the associated character vector are the column names to be resolved from dt}
}
\description{
Given a data.table and a list containing edges and their attributes, creates a data.table that can be used to create the edges in an igraph graph
}
\keyword{graph,}
\keyword{igraph}

