% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph_functions.R
\name{make_vertices_dt}
\alias{make_vertices_dt}
\title{Create a data.table that can be used to load all vertices into the graph}
\usage{
make_vertices_dt(dt, vertex_info, same_v_attr_list = NULL, parse_name = F)
}
\arguments{
\item{dt:}{data.table containing columns for all the vertexes and vertex attributes}

\item{vertex_info:}{named list, where the name is the vertex column name and the contents of each named entry is a character vector containing the names of the attribute columns to be associated with the vertex of interest}

\item{same_v_attr_list:}{named list where the name is the title to resolve the associated character vector names to}

\item{parse_name:}{whether or not to parse the entity_type and add it to the node name}
}
\description{
Given a data.table and a list containing vertices and their attributes, creates a data.table that can be used to create the vertices in an igraph graph
}
\examples{
Set information pertaining to which vertices and attributes will be added
vertex_info <- list(prchsr_unique = c('prchsr', 'purchase_pieces'),
                    candy = c('candy_typ', 'pieces'),
                    house = c('the_name')
same_v_attr_list <- list(all_pieces = c('purchase_pieces', 'pieces'))
}
\keyword{graph,}
\keyword{igraph}

