% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ip_functions.R
\name{mask}
\alias{mask}
\title{Get bits of the mask given the total bits}
\usage{
mask(bits, totalBits)
}
\arguments{
\item{bits:}{number of bits in mask you're looking for}

\item{totalBits:}{the total number of bits}
}
\description{
This function is a helper to maskIP below.
}
\keyword{ip,}
\keyword{mask}

