% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ip_functions.R
\name{maskIP}
\alias{maskIP}
\title{Converting an IP Address based on its mask}
\usage{
maskIP(ipAddr, maskLen, fieldBits=8, fieldSep=".")
}
\arguments{
\item{ipAddr:}{character form of the ip address to be converted}

\item{maskLen:}{integer indicating the mask length to which the ip address should be converted}

\item{fieldBits:}{integer indicating the number of bits in each field}

\item{fieldSep:}{character indicating what should be used to separate fields}
}
\description{
This function takes in an ip address and a mask length (number of bits) and converts the IP address to its form for the that mask length.
}
\examples{
maskIP ("192.7.16.42", 16)  #returns "192.7.0.0"
}
\keyword{ip,}
\keyword{mask}

