% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outlier_detection_functions.R
\name{outlier_ensemble}
\alias{outlier_ensemble}
\title{Master function to three outlier detection methods at once and ensemble the results}
\usage{
outlier_ensemble (input_df, dbscan_eps_range = seq(0.5, 2, 0.3), dbscan_MinPts_range = seq(3, 6, 1), dbscan_scale = T, kmeans_cluster_max = 10, kmeans_percentile = 95, kmeans_scale = T, cade_prop = 1, all_skip = c(), cade_range = seq(0.3, 0.7, 0.1))
}
\arguments{
\item{all:}{see documentation on the three outlier detection functions included in this package}
}
\description{
Given a data.frame of observations, determines outliers using dbscan, kmeans, and CADE and returns a dataframe indicating whether each observation was an outlier for that method as well as a sum of the number of the outlier detection methods that voted that observation as an outlier.
}

