% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph_functions.R
\name{retrieve_vertex_attributes}
\alias{retrieve_vertex_attributes}
\title{Retrieve all vertex attributes}
\usage{
retrieve_vertex_attributes(g, index = NULL)
}
\arguments{
\item{g:}{graph to use}

\item{index:}{numeric or named vector indicating the node(s) of interest}
}
\description{
Given vertex indices, retrieve all vertex attributes for those nodes
}
\examples{
retrieve_vertex_attributes(my_graph, 1:10)
}
\keyword{graph,}
\keyword{igraph}

