% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/propensity_matching_functions.R
\name{run_propensity_matching}
\alias{run_propensity_matching}
\title{Performs propensity score matching on a data.table and gives diagnostics}
\usage{
run_propensity_matching(dt, cols_of_interest, reg_outcome_var, match_exclude_cols, bad_entities, entity_name, plot_fp, list_save_fp, list_name, control_match_num = 3, verbose = T, sample_size = NULL, entity_sample = NULL)
}
\arguments{
\item{dt:}{data.table containin all the data}

\item{cols_of_interest:}{columns on which matching should occur}

\item{reg_outcome_var:}{character indicating the outcome column variable}

\item{match_exclude_cols:}{character vector of columns to exclude when performing the matching}

\item{bad_entities:}{a character vector indicating the entities for which propensity score matching should be performed}

\item{entity_name:}{character vector of the column in dt to be used subsetting down and identifiying 'bad_entities'}

\item{plot_fp:}{character filepath indicating where the psm matching plots should be stored}

\item{list_save_fp:}{character filepath indicating where the list of significant entities should be stored}

\item{control_match_num:}{The number of properties to be matched from the global population}

\item{verbose:}{Boolean indicating whether or not to give progress updates}

\item{sample_size:}{numeric indicating the sample size of the entity of interest to be taken for matching}
}
\description{
Given a data.table with information about which entities should be matched against, performs propensity score matching using nearest neighbors and saves out diagnostic plots. Returns a list of those entities which are significantly different from the population
}
\keyword{matching}
\keyword{propensity}
\keyword{score}

