% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ip_functions.R
\name{split_ip_network}
\alias{split_ip_network}
\title{Splitting a column containing IP address and mask into separate columns for each}
\usage{
split_ip_network(dt, net_name)
}
\arguments{
\item{dt:}{data.frame or data.table}

\item{net_name:}{character indicating the name of the column containing the ip address and mask}
}
\description{
This function takes a data.frame/table and a column name that indicates a column containing a combination of ip address and mask (e.g. '192.168.1.101/16') and splits this column into two separate columns. Returns a data.table.
}
\examples{
fake_dt <- data.table::data.table(ip = c('192.168.1.101/16', '10.10.2.201/24'), b = 1:2)
split_ip_network(fake_dt, 'ip')
}
\keyword{ip,}
\keyword{mask}

