% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph_functions.R
\name{subgraph_from_v_attr}
\alias{subgraph_from_v_attr}
\title{Creates a subgraph based on a vertex attribute}
\usage{
subgraph_from_v_attr(g, attr_values, attribute)
}
\arguments{
\item{g:}{graph to be subset}

\item{attr_values:}{Character vector of values of the vertex attribute that are allowed}

\item{attribute:}{Character vector of the vertex attribute upon which to subset}
}
\description{
Given a graph, desired attribute values, and the title of the attribute, this function creates a subgraph based on those criteria
}
\examples{
subgraph_from_v_attr(my_graph, c('yes', 'no'), 'is_it_good')
}
\keyword{graph,}
\keyword{igraph}

