% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/igraph_functions.R
\name{subset_graph_on_dates}
\alias{subset_graph_on_dates}
\title{subset the graph based on date ranges}
\usage{
subset_graph_on_dates(g, start_date, end_date, edge_date_var = "rel_date")
}
\description{
subset the graph based on date ranges
}

