% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{trim.leading}
\alias{trim.leading}
\title{Trim leading whitespace}
\usage{
trim.leading(x)
}
\arguments{
\item{x:}{character vector}
}
\description{
Given a character vector, trims all leading whitespace from each component of the vector
}
\examples{
trim.leading(c('  hello', 'no'))
}
\keyword{strings}

