% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{trim}
\alias{trim}
\title{Trim trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x:}{character vector}
}
\description{
Given a character vector, trims all leading and trailing whitespace from each component of the vector
}
\examples{
trim(c('hello  ', '  no'))
}
\keyword{strings}

