% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory_functions.R
\name{trim.trailing}
\alias{trim.trailing}
\title{Trim trailing whitespace}
\usage{
trim.trailing(x)
}
\arguments{
\item{x:}{character vector}
}
\description{
Given a character vector, trims all trailing whitespace from each component of the vector
}
\examples{
trim.trailing(c('hello  ', 'no'))
}
\keyword{strings}

